third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/xsl-unaffected-by-style-src-2-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
The text below should indicate that the style sheet was applied.
Style sheet applied.
